+++ packages/ant-design/src/__tests__/Table/faas.test.tsx
@@ -1,7 +1,6 @@
 /**
  * @jest-environment jsdom
  */
-import { FaasReactClient } from '@faasjs/react'
 import { render, screen } from '@testing-library/react'
 import { Table } from '../../Table'
 
@@ -18,7 +17,6 @@ describe('Table/faas', () => {
           text: async () => Promise.resolve('{"data":[{"test":"value"}]}'),
         }) as unknown as Promise<Response>
     )
-    FaasReactClient({ domain: '/' })
   })
 
   afterEach(() => {