+++ packages/react/src/__tests__/FaasDataWrapper.test.tsx
@@ -2,12 +2,7 @@
  * @jest-environment jsdom
  */
 import { render, screen } from '@testing-library/react'
-import {
-  FaasReactClient,
-  FaasDataWrapper,
-  type FaasDataInjection,
-  withFaasData,
-} from '..'
+import { FaasDataWrapper, type FaasDataInjection, withFaasData } from '..'
 import userEvent from '@testing-library/user-event'
 import { useState } from 'react'
 import { Response, setMock } from '@faasjs/browser'
@@ -29,8 +24,6 @@ describe('FaasDataWrapper', () => {
         })
       )
     })
-
-    FaasReactClient({ domain: '/' })
   })
 
   afterEach(() => {